perf: risk detail
