+++ src/router/pam/security.js
@@ -119,6 +119,16 @@ export default [
           title: i18n.t('AccountCheckUpdate'),
           permissions: ['accounts.view_accountrisk']
         }
+      },
+      {
+        path: ':id',
+        component: () => import('@/views/pam/RiskDetect/AccountCheckDetail/index.vue'),
+        name: 'AccountCheckDetail',
+        meta: {
+          title: i18n.t('RiskDetectionDetail'),
+          permissions: ['accounts.view_accountrisk']
+        },
+        hidden: true
       }
     ]
   }
